fix(ui): RF Gain format and Power QRP/QRO transitions
RF Gain:
- Fixed command format to RG-nn; and RG$-nn; (with minus sign)
- Range corrected to 0-60 (was incorrectly 0-100)
- Button faces display with minus sign (e.g., "-0", "-30")

Power:
- QRP (≤10W): 0.1W increments with PCnnnL; format
- QRO (>10W): 1W increments with PCnnnH; format
- Smooth transitions at 10W threshold
- Button face shows decimals for QRP, whole numbers for QRO

+++ src/ui/sidecontrolpanel.h
@@ -40,7 +40,7 @@ public:
     // Voice mode values
     void setMicGain(int gain);     // 0-80
     void setCompression(int comp); // 0-30
-    void setPower(int power);
+    void setPower(double power);
     void setDelay(double delay);
     void setBandwidth(double bw);
     void setHighCut(double hi); // kHz
